+++ R/cnd.R
@@ -261,8 +261,8 @@ cnd_call <- function(call) {
 #' and a full backtrace with `summary(last_error())`.
 #'
 #' You can also display a backtrace with the error message by setting
-#' the experimental option `rlang__backtrace_on_error`. It supports the
-#' following values:
+#' the option `rlang_backtrace_on_error`. It supports the following
+#' values:
 #'
 #' * `"reminder"`: Invite users to call `rlang::last_error()` to see a
 #'   backtrace.
@@ -689,8 +689,8 @@ conditionMessage.rlang_error <- function(c) {
 #' Errors thrown with [abort()] automatically save a backtrace that
 #' can be inspected by calling [last_error()]. Optionally, you can
 #' also display the backtrace alongside the error message by setting
-#' the experimental option `rlang__backtrace_on_error` to one of the
-#' following values:
+#' the option `rlang_backtrace_on_error` to one of the following
+#' values:
 #'
 #' * `"reminder"`: Display a reminder that the backtrace can be
 #'   inspected by calling [rlang::last_error()].
@@ -708,12 +708,12 @@ conditionMessage.rlang_error <- function(c) {
 #'   call [last_error()] to print the backtrace or inspect its data.
 #'
 #' * It prints the backtrace for the current error according to the
-#'   [`rlang__backtrace_on_error`] option.
+#'   [`rlang_backtrace_on_error`] option.
 #'
 #' This experimental handler is not exported because it is meant for
 #' interactive use only.
 #'
-#' @name rlang__backtrace_on_error
+#' @name rlang_backtrace_on_error
 #' @aliases add_backtrace
 #'
 #' @examples
@@ -721,7 +721,7 @@ conditionMessage.rlang_error <- function(c) {
 #' # errors:
 #'
 #' # options(
-#' #   rlang__backtrace_on_error = "branch",
+#' #   rlang_backtrace_on_error = "branch",
 #' #   error = rlang:::add_backtrace
 #' # )
 #' # stop("foo")
@@ -730,9 +730,10 @@ NULL
 format_onerror_backtrace <- function(trace) {
   show_trace <- show_trace_p()
 
-  if (!is_string(show_trace) || !show_trace %in% c("none", "reminder", "branch", "collapse", "full")) {
-    options(rlang__backtrace_on_error = NULL)
-    warn("Invalid `rlang__backtrace_on_error` option (resetting to `NULL`)")
+  opts <- c("none", "reminder", "branch", "collapse", "full")
+  if (!is_string(show_trace) || !show_trace %in% opts) {
+    options(rlang_backtrace_on_error = NULL)
+    warn("Invalid `rlang_backtrace_on_error` option (resetting to `NULL`)")
     return(NULL)
   }
 
@@ -775,7 +776,16 @@ format_onerror_backtrace <- function(trace) {
 }
 
 show_trace_p <- function() {
-  opt <- peek_option("rlang__backtrace_on_error")
+  old_opt <- peek_option("rlang__backtrace_on_error")
+  if (!is_null(old_opt)) {
+    warn_deprecated(paste_line(
+      "`rlang__backtrace_on_error` is no longer experimental.",
+      "It has been renamed to `rlang_backtrace_on_error`. Please update your RProfile."
+    ))
+    return(old_opt)
+  }
+
+  opt <- peek_option("rlang_backtrace_on_error")
   if (!is_null(opt)) {
     return(opt)
   }